Feature/chapter 11 supabase

+++ src/app/admin/categories/[id]/page.tsx
@@ -2,37 +2,36 @@
 
 import { useState, useEffect } from 'react'
 import { useRouter } from 'next/navigation'
-import type { Category, CategoryResponse, PostDetailsProps } from '@/app/_types/index'
+import type { Category, PostDetailsProps } from '@/app/_types/index'
 import CategoryForm from '@/app/admin/categories/_components/CategoryForm'
+import { useCategory } from '@/app/_hooks/useCategory'
+import { useSupabaseSession } from '@/app/_hooks/useSupabaseSession'
 
 const EditCategory: React.FC<PostDetailsProps> = ({ params }) => {
   const { id } = params
+  const { category: fetchedCat, isLoading, isError } = useCategory(id)
   const router = useRouter()
   const [category, setCategory] = useState<Category | null>(null)
-  const [loading, setLoading] = useState<boolean>(false)
+  const { token } = useSupabaseSession()
 
   useEffect(() => {
-    const fetcher = async () => {
-      setLoading(true)
-      const res: Response = await fetch(`${process.env.NEXT_PUBLIC_APP_BASE_URL}/api/admin/categories/${id}`)
-      const data = await res.json() as CategoryResponse
-      setCategory(data.category)
-      setLoading(false) 
-    }
-    fetcher()
-  },[id])
+    setCategory(fetchedCat ?? null)
+  },[fetchedCat])
 
   const handleEdit = async () => {
-    setLoading(true)
+    if (!token) return 
+
     const response: Response = await fetch(`${process.env.NEXT_PUBLIC_APP_BASE_URL}/api/admin/categories/${id}`, {
       method: 'PUT',
-      headers: {'Content-Type': 'application/json'},
+      headers: {
+        'Content-Type': 'application/json',
+        'Authorization': token,
+      },
       body: JSON.stringify({
         name: category?.name,
         posts: category?.posts.map(elem => ({id: elem.postId}))
       })
       })
-    setLoading(false)
 
     if (response.ok) {
       window.alert("更新しました")
@@ -43,11 +42,15 @@ const EditCategory: React.FC<PostDetailsProps> = ({ params }) => {
   }
 
   const handleDelete = async () => {
-    setLoading(true)
+    if (!token) return 
+    
     const response: Response = await fetch(`${process.env.NEXT_PUBLIC_APP_BASE_URL}/api/admin/categories/${id}`, {
       method: 'DELETE',
+      headers: {
+        'Content-Type': 'application/json',
+        'Authorization': token,
+      },
       })
-    setLoading(false)
 
     if (response.ok) {
       window.alert("削除しました")
@@ -57,8 +60,8 @@ const EditCategory: React.FC<PostDetailsProps> = ({ params }) => {
     }
   }
 
-  if (loading) return <p>読み込み中です...</p>
-  if (!category) return <div className="p-10 text-center text-3xl">404: カテゴリーが見つかりませんでした</div>
+  if (isLoading) return <p>読み込み中です...</p>
+  if (isError) return <div className="p-10 text-center text-3xl">404: カテゴリーが見つかりませんでした</div>
 
   return (
     <div className="p-8 bg-white w-full">
